requirements.txt
torch==2.0.1
wandb==0.15.3
transformers==4.41.2
torchaudio==2.0.2
tokenizers==0.19.1
pytorch-lightning==1.9.4
peft==0.9.0
librosa==0.10.1
jiwer==3.0.3
huggingface-hub==0.23.0
datasets==2.2.1
accelerate==0.30.0
streamlit==1.34.0
audio_recorder_streamlit==0.0.8
absl-py==2.1.0
gin-config==0.5.0